Split the pivot into its own component and make angle units configurable
The rotation pivot becomes a standalone PivotComponent, independent of
RotationComponent in both directions: an entity can rotate without one, and
can carry one with no rotation purely to anchor its scaling. TransformSystem
resolves the normalized anchor against the entity's own geometry inline,
which retires PivotResolveSystem, its extra pass, and the derived pivotOffset
field that could go stale behind a write.

Angles are still stored and composed in radians everywhere. The new AngleUnit
config key picks the unit they are authored and displayed in, converted at the
Lua and editor boundaries only.

Three behaviour changes, none of which fails loudly:

- AngleUnit defaults to degrees, so a project authored in radians must set
  AngleUnit=radians or every angle is reinterpreted. The effective unit is
  logged at startup.
- Scale now acts about the anchor, so a scaled entity holds its centre rather
  than its top-left corner. Anything at scale 1 is untouched. Author
  pivot = { x = 0, y = 0 } to restore corner-anchored growth per entity.
- A child carrying its own geometry under a rotated parent moves even at
  scale 1. Its anchor previously did not travel through the parent's rotation,
  so it span about the wrong point; Compose now rotates the anchor and backs
  the corner off by the scaled offset.

Also fixes PositionComponent.value, which was bound through a by-value
property: writes like pos.value.x = 5 mutated a temporary and silently did
nothing. It now writes through.

+++ src/Components/GlobalTransformComponent.h
@@ -12,6 +12,6 @@
 struct GlobalTransformComponent {
   glm::vec2 position{0.0f, 0.0f};
   glm::vec2 scale{1.0f, 1.0f};
-  double rotation{0.0};
+  float rotation{0.0F};
   glm::vec2 pivot{0.0f, 0.0f};
 };
